requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
Django==4.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-submodel==0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
fire==0.4.0
gunicorn==20.1.0
idna==3.3
PyJWT==2.4.0
python-dotenv==0.20.0
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
termcolor==1.1.0
urllib3==1.26.12
whitenoise==6.2.0
